refactor: atom폴더의 파일을 atoms폴더에 합침
