+++ src/pages/Landing/components/box.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import './box.scss';
-import { LocalImage } from '../../../common/atom/image';
-import { DescriptionContainer } from '../../../common/atom/container';
+import { LocalImage } from '../../../common/atoms/image';
+import { DescriptionContainer } from '../../../common/atoms/container';
 
 interface IconDescBoxProps {
     img: string;
